Feat: 문서, 파일 검색 시 폴더함 이동 버튼 구현
