+++ drive/src/main/java/com/Dolmeng_E/drive/domain/drive/dto/DocumentKafkaSaveDto.java
@@ -29,5 +29,7 @@ public class DocumentKafkaSaveDto {
         private LocalDateTime createdAt;
         private String rootType;
         private String rootId;
+        private String parentId;
+        private String fileUrl;
     }
 }
