+++ search/src/main/java/com/Dolmeng_E/search/domain/search/dto/EventDto.java
@@ -29,6 +29,8 @@ public class EventDto {
         private LocalDateTime createdAt;
         private String rootType;
         private String rootId;
+        private String parentId;
+        private String fileUrl;
     }
 }
