+++ search/src/main/java/com/Dolmeng_E/search/domain/search/service/FileEventConsumer.java
@@ -6,95 +6,25 @@ import com.Dolmeng_E.search.domain.search.entity.FileDocument;
 import com.Dolmeng_E.search.domain.search.repository.DocumentDocumentRepository;
 import com.Dolmeng_E.search.domain.search.repository.FileDocumentRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.redis.core.HashOperations;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Component;
 
-import javax.lang.model.UnknownEntityException;
-import java.time.format.DateTimeFormatter;
 import java.util.Map;
 import java.util.Optional;
 
 @Component
-public class EventConsumer {
+public class FileEventConsumer {
     private final ObjectMapper objectMapper; // JSON 파싱용
     private final HashOperations<String, String, String> hashOperations; // 유저 정보 가져오는 용도
-    private final DocumentDocumentRepository documentDocumentRepository;
     private final FileDocumentRepository fileDocumentRepository;
 
-    public EventConsumer(ObjectMapper objectMapper, RedisTemplate<String, String> redisTemplate, DocumentDocumentRepository documentDocumentRepository, FileDocumentRepository fileDocumentRepository) {
+    public FileEventConsumer(ObjectMapper objectMapper, RedisTemplate<String, String> redisTemplate, DocumentDocumentRepository documentDocumentRepository, FileDocumentRepository fileDocumentRepository) {
         this.objectMapper = objectMapper;
         this.hashOperations = redisTemplate.opsForHash();
-        this.documentDocumentRepository = documentDocumentRepository;
         this.fileDocumentRepository = fileDocumentRepository;
-    }
-
-    @KafkaListener(topics = "document-topic", groupId = "search-consumer-group")
-    public void handleDocument(String eventMessage, Acknowledgment ack) {
-        try {
-            // 1. Kafka 메시지(JSON)를 DTO로 파싱
-            EventDto eventDto = objectMapper.readValue(eventMessage, EventDto.class);
-            String eventType = eventDto.getEventType();
-            EventDto.EventPayload eventPayload = eventDto.getEventPayload();
-
-            // 2. eventType에 따라 분기 처리
-            switch (eventType) {
-                case "DOCUMENT_CREATED":
-                    // 생성
-                    String key = "user:"+eventPayload.getCreatedBy();
-                    Map<String, String> userInfo = hashOperations.entries(key);
-                    DocumentDocument document = DocumentDocument.builder()
-                            .id(eventPayload.getId())
-                            .docType("DOCUMENT")
-                            .searchTitle(eventPayload.getSearchTitle())
-                            .searchContent(eventPayload.getSearchContent())
-                            .dateTime(eventPayload.getCreatedAt().toLocalDate())
-                            .viewableUserIds(eventPayload.getViewableUserIds())
-                            .createdBy(eventPayload.getCreatedBy())
-                            .creatorName(userInfo.get("name"))
-                            .profileImageUrl(userInfo.get("profileImageUrl"))
-                            .rootId(eventPayload.getRootId())
-                            .rootType(eventPayload.getRootType())
-                            .build();
-                    documentDocumentRepository.save(document); // ES에 저장 또는 업데이트
-                    System.out.println("ES 색인(C/U) 성공: " + document.getId());
-                    ack.acknowledge();
-                    break;
-                case "DOCUMENT_UPDATED":
-                    Optional<DocumentDocument> optionalDocument = documentDocumentRepository.findById(eventPayload.getId());
-                    if (optionalDocument.isPresent()) {
-                        DocumentDocument documentToUpdate = optionalDocument.get();
-                        documentToUpdate.updateDocument(eventPayload);
-                        documentDocumentRepository.save(documentToUpdate);
-                        System.out.println("ES 업데이트(U) 성공: " + documentToUpdate.getId());
-                    } else {
-                        System.err.println("ES 업데이트(U) 실패: 원본 문서를 찾을 수 없음 - ID: " + eventPayload.getId());
-                    }
-                    ack.acknowledge();
-                    break;
-
-                case "DOCUMENT_DELETED":
-                    // 삭제는 ID만 파싱
-                    String deletedDocumentId = eventPayload.getId();
-
-                    documentDocumentRepository.deleteById(deletedDocumentId); // ES에서 삭제
-                    ack.acknowledge();
-                    System.out.println("ES 삭제(D) 성공: " + deletedDocumentId);
-                    break;
-
-                default:
-                    // 알 수 없는 이벤트 타입 로그 처리
-                    ack.acknowledge();
-                    System.err.println("알 수 없는 이벤트 타입: " + eventType);
-                    break;
-            }
-
-        } catch (Exception e) {
-            System.err.println("ES 처리 실패: " + e.getMessage());
-        }
     }
 
     @KafkaListener(topics = "file-topic", groupId = "search-consumer-group")
@@ -123,6 +53,8 @@ public class EventConsumer {
                             .profileImageUrl(userInfo.get("profileImageUrl"))
                             .rootId(eventPayload.getRootId())
                             .rootType(eventPayload.getRootType())
+                            .parentId(eventPayload.getParentId())
+                            .fileUrl(eventPayload.getFileUrl())
                             .build();
                     fileDocumentRepository.save(fileDocument); // ES에 저장 또는 업데이트
                     System.out.println("ES 색인(C/U) 성공: " + fileDocument.getId());
@@ -132,7 +64,10 @@ public class EventConsumer {
                     Optional<FileDocument> optionalFile = fileDocumentRepository.findById(eventPayload.getId());
                     if (optionalFile.isPresent()) {
                         FileDocument documentToUpdate = optionalFile.get();
-                        documentToUpdate.updateFile(eventPayload);
+                        if(eventPayload.getSearchTitle()!=null){
+                            documentToUpdate.updateFile(eventPayload);
+                        }
+                        documentToUpdate.setParentId(eventPayload.getParentId());
                         fileDocumentRepository.save(documentToUpdate);
                         System.out.println("ES 업데이트(U) 성공: " + documentToUpdate.getId());
                     } else {
